FEATURE: rename page -> content
