+++ src/quant/contents/__init__.py
@@ -1,8 +1,8 @@
-from pages.backtesting import BackTestingManager
-from pages.login import LoginManager
+from contents.backtesting import BackTestingManager
+from contents.login import LoginManager
 
 # display order
-pages = {
+contents = {
     LoginManager.name: LoginManager(),
     BackTestingManager.name: BackTestingManager(),
 }
