+++ src/quant/contents/cache.py
