+++ src/quant/dashboard.py
@@ -1,18 +1,18 @@
 import streamlit as st
 from streamlit_option_menu import option_menu
 
-from pages import pages
+from contents import contents
 
 st.set_page_config(layout="wide")
 
 with st.sidebar:
-    selected_page = option_menu(
+    selected_content = option_menu(
         menu_title="Contents",  # required
-        options=[page.name for page in pages.values()],
-        icons=[page.icon for page in pages.values()],
+        options=[content.name for content in contents.values()],
+        icons=[content.icon for content in contents.values()],
         menu_icon="cast",
         default_index=1,
     )
 
 # display info
-pages[selected_page].run()
+contents[selected_content].run()
